requirements.txt
openai>=1.30.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
pydantic>=2.6.0
